+++ src/Settings.ts
@@ -602,7 +602,7 @@ export const DEFAULT_SETTINGS: CodeStylerSettings = {
 	excludedLanguages: EXCLUDED_LANGUAGES,
 	processedCodeblocksWhitelist: WHITELIST_CODEBLOCKS,
 	redirectLanguages: {},
-	version: "1.1.3",
+	version: "1.1.4",
 };
 
 export function convertSettings(settings: CodeStylerSettings): CodeStylerSettings {
@@ -660,6 +660,7 @@ const settingsUpdaters: Record<string,(settings: CodeStylerSettings)=>CodeStyler
 	"1.1.1": settingsPreserve,
 	"1.1.2": settingsPreserve,
 	"1.1.3": settingsPreserve,
+	"1.1.4": settingsPreserve,
 };
 
 // Constants